+++ src/Data/StaticText.hs
@@ -12,9 +12,11 @@
 
 {-|
 
-Sext (/s/tatic t/ext/) provides type-level safety for basic operations
-on string-like types (finite lists of elements). Use it when you need
-static guarantee on lengths of strings produced in your code.
+static-text provides type-level safety for basic operations on
+string-like types (finite lists of elements), such as "Data.Text",
+"String" (and all lists), "Data.ByteString" and "Data.Vector". Use it
+when you need static guarantee on lengths of strings produced in your
+code.
 
 An example application would be a network exchange protocol built of
 packets with fixed-width fields:
@@ -25,9 +27,9 @@ packets with fixed-width fields:
 {\-\# LANGUAGE TemplateHaskell #-\}
 @
 
-> import Data.Sext
+> import Data.StaticText
 >
-> mkPacket :: ByteString -> Sext 32 ByteString
+> mkPacket :: ByteString -> StaticText 32 ByteString
 > mkPacket inp =
 >   -- 5-character version signature
 >   $(sext "PKT10") `append`
@@ -37,25 +39,26 @@ packets with fixed-width fields:
 >   checksum
 >   where
 >     payload = createLeft 0x20 inp
->     checksum :: Sext 2 ByteString
+>     checksum :: StaticText 2 ByteString
 >     checksum = createLeft 0x20 $
->                pack $ show $ Data.Sext.length payload `mod` 100
+>                pack $ show $ Data.StaticText.length payload `mod` 100
 >
-> message :: Sext 64 ByteString
+> message :: StaticText 64 ByteString
 > message = mkPacket "Hello" `append` mkPacket "world"
 
-Sext combinators are defined for members of 'Sextable' class. The
-package includes 'Sextable' instances for several common types.
+static-text combinators are defined for members of 'IsStaticText'
+class. The package includes 'IsStaticText' instances for several
+common types.
 
 This module is meant to be imported qualifed, e.g.
 
-> import qualified Data.Sext as S
+> import qualified Data.StaticText as S
 
 -}
 
-module Data.Sext
+module Data.StaticText
        (
-         -- * Constructing Sexts
+         -- * Constructing StaticTexts
          --
          -- | See also 'C.unsafeCreate'
          createLeft
@@ -64,7 +67,7 @@ module Data.Sext
        , create
        , replicate
 
-         -- * Working with Sexts
+         -- * Working with StaticTexts
        , append
        , take
        , drop
@@ -74,9 +77,9 @@ module Data.Sext
 
        , length
 
-         -- * Sextable class
-       , Sext
-       , Sextable(Elem, unsafeCreate, unwrap)
+         -- * IsStaticText class
+       , StaticText
+       , IsStaticText(Elem, unsafeCreate, unwrap)
        )
 
 where
@@ -86,9 +89,9 @@ import           Prelude as P hiding (drop, length, map, replicate, take)
 import           GHC.TypeLits
 
 import           Data.Proxy
-import           Data.Sext.Class (Elem, Sext, Sextable)
-import qualified Data.Sext.Class as C
-import           Data.Sext.TH
+import           Data.StaticText.Class (Elem, StaticText, IsStaticText)
+import qualified Data.StaticText.Class as C
+import           Data.StaticText.TH
 
 
 -- $setup
@@ -98,19 +101,19 @@ import           Data.Sext.TH
 -- >>> import Data.Char (toUpper)
 
 
--- | Safely create a Sext, possibly altering the source to match
+-- | Safely create a StaticText, possibly altering the source to match
 -- target length. If target length is less than that of the source,
 -- the source gets truncated. If target length is greater, the source
 -- is padded using the provided basic element. Elements on the left
 -- are preferred.
 --
--- >>> createLeft ' ' "foobarbaz" :: Sext 6 String
+-- >>> createLeft ' ' "foobarbaz" :: StaticText 6 String
 -- "foobar"
--- >>> createLeft '#' "foobarbaz" :: Sext 12 String
+-- >>> createLeft '#' "foobarbaz" :: StaticText 12 String
 -- "foobarbaz###"
 createLeft :: forall a i.
-              (Sextable a, KnownNat i) =>
-              Elem a -> a -> Sext i a
+              (IsStaticText a, KnownNat i) =>
+              Elem a -> a -> StaticText i a
 createLeft e s =
   C.unsafeCreate $
   C.take t $
@@ -122,13 +125,13 @@ createLeft e s =
 
 -- | Just like 'createLeft', except that elements on the right are preferred.
 --
--- >>> createRight '@' "foobarbaz" :: Sext 6 String
+-- >>> createRight '@' "foobarbaz" :: StaticText 6 String
 -- "barbaz"
--- >>> createRight '!' "foobarbaz" :: Sext 12 String
+-- >>> createRight '!' "foobarbaz" :: StaticText 12 String
 -- "!!!foobarbaz"
 createRight :: forall a i.
-               (Sextable a, KnownNat i) =>
-               Elem a -> a -> Sext i a
+               (IsStaticText a, KnownNat i) =>
+               Elem a -> a -> StaticText i a
 createRight e s =
   C.unsafeCreate $
   C.drop (C.length s - t) $
@@ -137,19 +140,19 @@ createRight e s =
     t = fromIntegral $ natVal (Proxy :: Proxy i)
 
 
--- | Attempt to safely create a Sext if it matches target length.
+-- | Attempt to safely create a StaticText if it matches target length.
 --
--- >>> create "foobar" :: Maybe (Sext 6 String)
+-- >>> create "foobar" :: Maybe (StaticText 6 String)
 -- Just "foobar"
--- >>> create "barbaz" :: Maybe (Sext 8 String)
+-- >>> create "barbaz" :: Maybe (StaticText 8 String)
 -- Nothing
 --
 -- This is safer than 'C.unsafeCreate' and unlike with 'createLeft' /
 -- 'createRight' the source value is left unchanged. However, this
 -- implies a further run-time check for Nothing values.
 create :: forall a i.
-          (Sextable a, KnownNat i) =>
-          a -> P.Maybe (Sext i a)
+          (IsStaticText a, KnownNat i) =>
+          a -> P.Maybe (StaticText i a)
 create s =
   if C.length s == t
   then Just $ C.unsafeCreate s
@@ -158,57 +161,57 @@ create s =
     t = fromIntegral $ natVal (Proxy :: Proxy i)
 
 
--- | Append two Sexts together.
+-- | Append two StaticTexts together.
 --
--- >>> append $(sext "foo") $(sext "bar") :: Sext 6 String
+-- >>> append $(sext "foo") $(sext "bar") :: StaticText 6 String
 -- "foobar"
 append :: forall a m n.
-          (Sextable a) => Sext m a -> Sext n a -> Sext (m + n) a
+          (IsStaticText a) => StaticText m a -> StaticText n a -> StaticText (m + n) a
 append a b = C.unsafeCreate $ C.append (C.unwrap a) (C.unwrap b)
 
 
--- | Construct a new Sext from a basic element.
+-- | Construct a new StaticText from a basic element.
 --
--- >>> replicate '=' :: Sext 10 String
+-- >>> replicate '=' :: StaticText 10 String
 -- "=========="
 replicate :: forall a i.
-             (Sextable a, KnownNat i) => Elem a -> Sext i a
+             (IsStaticText a, KnownNat i) => Elem a -> StaticText i a
 replicate e =
   C.unsafeCreate $ C.replicate t e
   where
     t = fromIntegral $ natVal (Proxy :: Proxy i)
 
 
--- | Map a Sext to a Sext of the same length.
+-- | Map a StaticText to a StaticText of the same length.
 --
--- >>> map toUpper $(sext "Hello") :: Sext 5 String
+-- >>> map toUpper $(sext "Hello") :: StaticText 5 String
 -- "HELLO"
-map :: Sextable a =>
-       (Elem a -> Elem a) -> Sext m a -> Sext m a
+map :: IsStaticText a =>
+       (Elem a -> Elem a) -> StaticText m a -> StaticText m a
 map f s =
   C.unsafeCreate $ C.map f $ C.unwrap s
 
 
--- | Reduce Sext length, preferring elements on the left.
+-- | Reduce StaticText length, preferring elements on the left.
 --
--- >>> take $(sext "Foobar") :: Sext 3 String
+-- >>> take $(sext "Foobar") :: StaticText 3 String
 -- "Foo"
 take :: forall a m n.
-        (Sextable a, KnownNat m, KnownNat n, n <= m) =>
-        Sext m a -> Sext n a
+        (IsStaticText a, KnownNat m, KnownNat n, n <= m) =>
+        StaticText m a -> StaticText n a
 take s =
   C.unsafeCreate $ C.take t $ C.unwrap s
   where
     t = fromIntegral $ natVal (Proxy :: Proxy n)
 
 
--- | Reduce Sext length, preferring elements on the right.
+-- | Reduce StaticText length, preferring elements on the right.
 --
--- >>> drop $(sext "Foobar") :: Sext 2 String
+-- >>> drop $(sext "Foobar") :: StaticText 2 String
 -- "ar"
 drop :: forall a m n.
-        (Sextable a, KnownNat m, KnownNat n, n <= m) =>
-        Sext m a -> Sext n a
+        (IsStaticText a, KnownNat m, KnownNat n, n <= m) =>
+        StaticText m a -> StaticText n a
 drop s =
   C.unsafeCreate $ C.drop (C.length s' - t) s'
   where
@@ -218,22 +221,22 @@ drop s =
 
 -- | Obtain value-level length.
 length :: forall a m.
-          KnownNat m => Sext m a -> P.Int
+          KnownNat m => StaticText m a -> P.Int
 length _ = P.fromIntegral P.$ natVal (Proxy :: Proxy m)
 
 
--- | Fill a Sext with extra elements up to target length, padding
+-- | Fill a StaticText with extra elements up to target length, padding
 -- original elements to the left.
 padLeft :: forall a m n.
-           (Sextable a, KnownNat m, KnownNat (n - m),
+           (IsStaticText a, KnownNat m, KnownNat (n - m),
             n ~ (n - m + m), m <= n) =>
-           Elem a -> Sext m a -> Sext n a
+           Elem a -> StaticText m a -> StaticText n a
 padLeft pad = append (replicate pad)
 
 
 -- | Like 'padLeft', but original elements are padded to the right.
 padRight :: forall a m n.
-           (Sextable a, KnownNat m, KnownNat (n - m),
+           (IsStaticText a, KnownNat m, KnownNat (n - m),
             n ~ (m + (n - m)), m <= n) =>
-           Elem a -> Sext m a -> Sext n a
+           Elem a -> StaticText m a -> StaticText n a
 padRight pad = P.flip append (replicate pad)